+++ packages/brick_offline_first/lib/brick_offline_first.dart
@@ -6,3 +6,4 @@ export 'package:brick_offline_first/src/offline_first_policy.dart';
 export 'package:brick_offline_first/src/annotations/offline_first_annotation.dart';
 export 'package:brick_offline_first/src/models/offline_first_model.dart';
 export 'package:brick_offline_first/src/models/offline_first_serdes.dart';
+export 'package:brick_offline_first/src/runtime_offline_first_definition.dart';