+++ packages/brick_offline_first_with_graphql_build/test/offline_first_generator/test_graphql_config_query_operation_transformer.dart
@@ -44,6 +44,9 @@ class QueryOperationTransformerExampleAdapter
   final queryOperationTransformer =
       QueryOperationTransformerExampleTransformer.new;
   @override
+  final fieldsToOfflineFirstRuntimeDefinition =
+      <String, RuntimeOfflineFirstDefinition>{};
+  @override
   final fieldsToGraphqlRuntimeDefinition = <String, RuntimeGraphqlDefinition>{
     'name': const RuntimeGraphqlDefinition(
       association: false,