+++ packages/brick_offline_first_with_graphql_build/test/offline_first_generator/test_offline_first_where_rename.dart
@@ -22,7 +22,7 @@ Future<Map<String, dynamic>> _$GraphqlConfigEndpointToGraphql(
     GraphqlConfigEndpoint instance,
     {required GraphqlProvider provider,
     OfflineFirstRepository? repository}) async {
-  return {'someField': instance.someField.name};
+  return {'name': instance.someField.name};
 }
 
 Future<GraphqlConfigEndpoint> _$GraphqlConfigEndpointFromSqlite(
@@ -53,6 +53,13 @@ class GraphqlConfigEndpointAdapter
     extends OfflineFirstAdapter<GraphqlConfigEndpoint> {
   GraphqlConfigEndpointAdapter();
 
+  @override
+  final fieldsToOfflineFirstRuntimeDefinition =
+      <String, RuntimeOfflineFirstDefinition>{
+    'someField': const RuntimeOfflineFirstDefinition(
+      where: <String, String>{'name': "data['name']"},
+    )
+  };
   @override
   final fieldsToGraphqlRuntimeDefinition = <String, RuntimeGraphqlDefinition>{
     'someField': const RuntimeGraphqlDefinition(